fix: 회원 이벤트 생성 버그 수정
