+++ server/src/main/java/haengdong/common/auth/infrastructure/AdminInterceptor.java
@@ -1,17 +1,19 @@
 package haengdong.common.auth.infrastructure;
 
+import haengdong.common.auth.application.AuthService;
+import haengdong.common.exception.AuthenticationException;
+import haengdong.common.exception.HaengdongErrorCode;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpMethod;
 import org.springframework.web.servlet.HandlerInterceptor;
-import haengdong.common.auth.application.AuthService;
-import haengdong.common.exception.AuthenticationException;
-import haengdong.common.exception.HaengdongErrorCode;
 
 @Slf4j
+@RequiredArgsConstructor
 public class AdminInterceptor implements HandlerInterceptor {
 
     public static final String LOGIN_MEMBER_REQUEST = "loginUserId";
@@ -22,11 +24,6 @@ public class AdminInterceptor implements HandlerInterceptor {
 
     private final AuthService authService;
     private final AuthenticationExtractor authenticationExtractor;
-
-    public AdminInterceptor(AuthService authService, AuthenticationExtractor authenticationExtractor) {
-        this.authService = authService;
-        this.authenticationExtractor = authenticationExtractor;
-    }
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
@@ -39,8 +36,8 @@ public class AdminInterceptor implements HandlerInterceptor {
     }
 
     private void validateToken(HttpServletRequest request) {
-        String token = authenticationExtractor.extract(request, authService.getTokenName());
-        Long userId = authService.findUserIdByToken(token);
+        String jwt = authenticationExtractor.extract(request, authService.getTokenName());
+        Long userId = authService.findUserIdByJWT(jwt);
         String uri = request.getRequestURI();
         Matcher matcher = ADMIN_URI_PATTERN.matcher(uri);
         if (!matcher.find()) {
