+++ server/src/main/java/haengdong/common/auth/infrastructure/AuthenticationPrincipalArgumentResolver.java
@@ -1,16 +1,20 @@
 package haengdong.common.auth.infrastructure;
 
-import static haengdong.common.auth.infrastructure.AdminInterceptor.LOGIN_MEMBER_REQUEST;
-
 import haengdong.common.auth.Login;
+import haengdong.common.auth.application.AuthService;
 import jakarta.servlet.http.HttpServletRequest;
+import lombok.RequiredArgsConstructor;
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 
+@RequiredArgsConstructor
 public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArgumentResolver {
+
+    private final AuthService authService;
+    private final AuthenticationExtractor authenticationExtractor;
 
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
@@ -20,7 +24,8 @@ public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArg
     @Override
     public Long resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
         HttpServletRequest request = webRequest.getNativeRequest(HttpServletRequest.class);
+        String jwt = authenticationExtractor.extract(request, authService.getTokenName());
 
-        return (Long) request.getAttribute(LOGIN_MEMBER_REQUEST);
+        return authService.findUserIdByJWT(jwt);
     }
 }
